+++ src/main/java/com/snack/news/controller/AdminController.java
@@ -3,6 +3,7 @@ package com.snack.news.controller;
 import com.snack.news.domain.news.News;
 import com.snack.news.dto.NewsDto;
 import com.snack.news.dto.WrappedResponse;
+import com.snack.news.service.AdminService;
 import com.snack.news.service.NewsService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -17,7 +18,12 @@ import java.util.Optional;
 @RequestMapping("admin/api")
 public class AdminController {
 
-	private final NewsService newsService;
+	private final AdminService adminService;
+
+	@PostMapping("/news")
+	public ResponseEntity<NewsDto> createNews(@Valid @RequestBody NewsDto newsDto) {
+		return WrappedResponse.ok(adminService.createNews(newsDto));
+	}
 
 	@GetMapping("/news")
 	public ResponseEntity<Page<News>> getNewsList() {
@@ -26,15 +32,10 @@ public class AdminController {
 
 	@GetMapping("/news/{page}")
 	public ResponseEntity<Page<News>> getNewsList(@PathVariable int page) {
-		Page<News> result = newsService.getNewsListForAdmin(page);
+		Page<News> result = adminService.getNewsList(page);
 		if (result.isEmpty()) {
 			return ResponseEntity.noContent().build();
 		}
 		return WrappedResponse.ok(result);
-	}
-
-	@PostMapping("/news")
-	public ResponseEntity<NewsDto> createNews(@Valid @RequestBody NewsDto newsDto) {
-		return WrappedResponse.ok(newsService.createNews(newsDto));
 	}
 }
